intentcp-core/pyproject.toml
[project]
name = "intentcp-core"
version = "0.1.0"
description = "IntentCP core server (FastAPI-based orchestration layer)"
authors = [
  { name = "Sim Jaebin", email = "simjabin@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "typer>=0.12.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  "tuya-iot-py-sdk>=0.6.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
]

[project.scripts]
intentcp = "intentcp_core.cli.main:app"

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]